use pytest for fast allreduce test
